Moved all files that made up sjljs core into sjl.js.  Moved all core classes into 'stdlib'.
