--- src/sjl/sjl-package.js
@@ -1,10 +1,0 @@
-/**
- * Created by Ely on 8/15/2015.
- */
-(function (context) {
-
-    'use strict';
-
-    context.sjl.createTopLevelPackage(context.sjl);
-
-}(typeof window === 'undefined' ? global : window));
