--- src/sjl/sjl-set-functions.js
@@ -1,309 +1,0 @@
-/**
- * Created by Ely on 5/24/2014.
- * Cartesian functions copied from "Javascript the definitive guide"
- * getValueFromObj and setValueOnObj are not from "Javascript ..."
- * @note using legacy getters and setters from within `extend` method requires a refactor
- * as it does not work with the deep option and should.
- */
-
-'use strict';
-
-(function (context) {
-
-    /**
-     * Used by sjl.extend definition
-     * @type {Function}
-     */
-    var getOwnPropertyDescriptor =
-            typeof Object.getOwnPropertyDescriptor === 'function'
-                ? Object.getOwnPropertyDescriptor : null,
-        sjl = context.sjl;
-
-    /**
-     * Checks if object has method key passed.
-     * @function module:sjl.hasMethod
-     * @param obj {Object|*} - Object to search on.
-     * @param method - Method name to search for.
-     * @returns {Boolean}
-     */
-    sjl.hasMethod = function (obj, method) {
-        return !sjl.isEmptyObjKeyOrNotOfType(obj, method, 'Function');
-    };
-
-    /**
-     * Searches obj for key and returns it's value.  If value is a function
-     * calls function if `raw` is set to `false`, with optional `args`, and returns it's return value.
-     * If `raw` is true returns the actual function if value found is a function.
-     * @function module:sjl.getValueFromObj
-     * @param key {String} The hash key to search for
-     * @param obj {Object} the hash to search within
-     * @param args {Array} optional the array to pass to value if it is a function
-     * @param raw {Boolean} optional whether to return value even if it is a function.  Default `true`.
-     * @todo allow this function to use getter function for key if it exists
-     * @param noLegacyGetters {Boolean} - Default false (use legacy getters).
-     *  Whether to use legacy getters to fetch the value ( get{key}() or overloaded {key}() )
-     *
-     * @returns {*}
-     */
-    sjl.getValueFromObj = function (key, obj, args, raw, noLegacyGetters) {
-        // Warn user(s) of new update to this function where `raw` is not being passed in.
-        //if (typeof raw === 'undefined') {
-        //    console.warn('`sjl.getValueFromObj` now has it\'s `raw` parameter set to `true` by default.  ' +
-        //        'This warning will be removed in the next library update.');
-        //}
-        args = args || null;
-        raw = raw || true;
-        noLegacyGetters = typeof noLegacyGetters === 'undefined' ? false : noLegacyGetters;
-
-        // Get qualified getter function names
-        var overloadedGetterFunc = sjl.camelCase(key, false),
-            getterFunc = 'get' + sjl.camelCase(key, true),
-            retVal = null;
-
-        // Resolve return value
-        if (key.indexOf('.') !== -1) {
-            retVal = sjl.namespace(key, obj);
-        }
-        // If obj has a getter function for key, call it
-        else if (!noLegacyGetters && sjl.hasMethod(obj, getterFunc)) {
-            retVal = obj[getterFunc]();
-        }
-        else if (!noLegacyGetters && sjl.hasMethod(obj, overloadedGetterFunc)) {
-            retVal = obj[overloadedGetterFunc]();
-        }
-        else if (typeof obj[key] !== 'undefined') {
-            retVal = obj[key];
-        }
-
-        // Decide what to do if return value is a function
-        if (sjl.classOfIs(retVal, 'Function') && sjl.empty(raw)) {
-            retVal = args ? retVal.apply(obj, args) : retVal.apply(obj);
-        }
-
-        // Return result of setting value on obj, else return obj
-        return retVal;
-    };
-
-    /**
-     * Sets a key to value on obj.
-     * @function module:sjl.setValueOnObj
-     * @param key {String} - Key to search for (can be a dot
-     * separated string 'all.your.base' will traverse {all: {your: {base: {...}}})
-     * @param value {*} - Value to set on obj
-     * @param obj {Object} - Object to set key to value on
-     * @returns {*|Object} returns result of setting key to value on obj or obj
-     * if no value resulting from set operation
-     */
-    sjl.setValueOnObj = function (key, value, obj) {
-        // Get qualified setter function name
-        var overloadedSetterFunc = sjl.camelCase(key, false),
-            setterFunc = 'set' + sjl.camelCase(key, true),
-            retVal = obj;
-
-        // Else set the value on the obj
-        if (key.indexOf('.') !== -1) {
-            retVal = sjl.namespace(key, obj, value);
-        }
-        // If obj has a setter function for key, call it
-        else if (sjl.hasMethod(obj, setterFunc)) {
-            retVal = obj[setterFunc](value);
-        }
-        else if (sjl.hasMethod(obj, overloadedSetterFunc)) {
-            retVal = obj[overloadedSetterFunc](value);
-        }
-        else {
-            obj[key] = typeof value !== 'undefined' ? value : null;
-        }
-
-        // Return result of setting value on obj, else return obj
-        return retVal;
-    };
-
-    /**
-     * Copy the enumerable properties of p to o, and return o.
-     * If o and p have a property by the same name, o's property is overwritten.
-     * This function does not handle getters and setters or copy attributes but
-     * does search for setter methods in the format "setPropertyName" and uses them
-     * if they are available for property `useLegacyGettersAndSetters` is set to true.
-     * @param o {mixed} - *object to extend
-     * @param p {mixed} - *object to extend from
-     * @param deep {Boolean} - Whether or not to do a deep extend (run extend on each prop if prop value is of type 'Object')
-     * @param useLegacyGettersAndSetters {Boolean} - Whether or not to do a deep extend (run extend on each prop if prop value is of type 'Object')
-     * @returns {*} - returns o
-     */
-     function extend (o, p, deep, useLegacyGettersAndSetters) {
-        deep = deep || false;
-        useLegacyGettersAndSetters = useLegacyGettersAndSetters || false;
-
-        var prop, propDescription,
-            classOf_p_prop,
-            classOf_o_prop;
-
-        // If `o` or `p` are not set bail
-        if (!sjl.isset(o) || !sjl.isset(p)) {
-            return o;
-        }
-
-        for (prop in p) { // For all props in p.
-            classOf_p_prop = sjl.issetObjKey(p, prop) ? sjl.classOf(p[prop]) : 'Empty';
-            classOf_o_prop = sjl.issetObjKey(o, prop) ? sjl.classOf(o[prop]) : 'Empty';
-
-            // If property is present on target (o) and is not writable, skip iteration
-            if (getOwnPropertyDescriptor) {
-                propDescription = getOwnPropertyDescriptor(o, prop);
-                if (propDescription && !propDescription.writable) {
-                    continue;
-                }
-            }
-
-            if (deep) {
-                if (classOf_o_prop === 'Object'
-                    && classOf_p_prop === 'Object'
-                    && !sjl.isEmptyObj(p[prop])) {
-                    extend(o[prop], p[prop], deep);
-                }
-                else if (useLegacyGettersAndSetters) {
-                    sjl.setValueOnObj(prop, sjl.getValueFromObj(prop, p, null, useLegacyGettersAndSetters), o);
-                }
-                else {
-                    o[prop] = p[prop];
-                }
-            }
-
-            // Else set
-            else {
-                o[prop] = p[prop];
-            }
-        }
-
-        return o;
-    }
-
-    /**
-     * Extends first object passed in with all other object passed in after.
-     * First param could be a boolean indicating whether or not to perform a deep extend.
-     * Last param could also be a boolean indicating whether to use legacy setters if they are available
-     * when extending one object with another.
-     *
-     * @example
-     *  var o = {setGreeting: v => this.greeting = 'Hello ' + v},
-     *      otherObject = {greeting: 'Junior'};
-     *  // Calls o.setGreeting when merging otherObject because `true` was passed in
-     *  // as the last parameter
-     *  sjl.extend(o, otherObject, true);
-     *
-     * @function module:sjl.extend
-     * @param [, Boolean, obj] {Object|Boolean} - If boolean, causes `extend` to perform a deep extend.  Optional.
-     * @param [, obj, obj] {Object} - Objects to hierarchically extend.
-     * @param [, Boolean] {Boolean} - Optional.
-     * @returns {Object} - Returns first object passed in.
-     */
-    sjl.extend = function () {
-        // Return if no arguments
-        if (arguments.length === 0) {
-            return;
-        }
-
-        var args = sjl.argsToArray(arguments),
-            deep = sjl.extractBoolFromArrayStart(args),
-            useLegacyGettersAndSetters = sjl.extractBoolFromArrayEnd(args),// Can't remove this until version 0.5 cause it may cause breaking changes in dependant projects
-            arg0 = args.shift();
-
-        // Extend object `0` with other objects
-        sjl.forEach(args, function (arg) {
-            // Extend `arg0` if `arg` is an object
-            if (sjl.classOfIs(arg, 'Object')) {
-                extend(arg0, arg, deep, useLegacyGettersAndSetters);
-            }
-        });
-
-        return arg0;
-    };
-
-    /**
-     * Returns copy of object.
-     * @function module:sjl.clone
-     * @param obj {Object}
-     * @returns {*} - Cloned object.
-     */
-    sjl.clone = function (obj) {
-        return  sjl.extend(true, {}, obj);
-    };
-
-    /**
-     * Returns copy of object using JSON stringify/parse.
-     * @function module:sjl.jsonClone
-     * @param obj {Object} - Object to clone.
-     * @returns {*} - Cloned object.
-     */
-    sjl.jsonClone = function (obj) {
-        return JSON.parse(JSON.stringify(obj));
-    };
-
-//    if (typeof sjl.merge === 'undefined') {
-        /**
-         * Copy the enumerable properties of p to o, and return o.
-         * If o and p have a property by the same name, o's property is left alone.
-         * This function does not handle getters and setters or copy attributes.
-         * @param o {mixed} - *object to merge to
-         * @param p {mixed} - *object to merge from
-         * @returns {*} - returns o
-         */
-//        sjl.merge = function (o, p) {
-//            for (prop in p) { // For all props in p.
-//                if (o.hasOwnProperty[prop]) continue; // Except those already in o.
-//                o[prop] = p[prop]; // Add the property to o.
-//            }
-//            return o;
-//        };
-//    }
-
-//    if (typeof sjl.subtract === 'undefined') {
-        /**
-         * For each property of p, delete the property with the same name from o.
-         * Return o.
-         */
-//        sjl.subtract = function (o, p) {
-//            for (prop in p) { // For all props in p
-//                delete o[prop]; // Delete from o (deleting a
-//                // nonexistent prop is harmless)
-//            }
-//            return o;
-//        };
-//    }
-
-//    if (typeof sjl.restrict === 'undefined') {
-        /**
-         * Remove properties from o if there is not a property with the same name in p.
-         * Return o.
-         */
-//        sjl.restrict = function (o, p) {
-//            for (prop in o) { // For all props in o
-//                if (!(prop in p)) delete o[prop]; // Delete if not in p
-//            }
-//            return o;
-//        };
-//    }
-
-//    if (typeof sjl.union === 'undefined') {
-        /**
-         * Return a new object that holds the properties of both o and p.
-         * If o and p have properties by the same name, the values from p are used.
-         */
-//        sjl.union = function (o, p, deep, useLegacyGettersAndSetters) {
-//            return sjl.extend(deep, sjl.clone(o), p, useLegacyGettersAndSetters);
-//        };
-//    }
-
-//    if (typeof sjl.intersection === 'undefined') {
-        /**
-         * Return a new object that holds only the properties of o that also appear
-         * in p. This is something like the intersection of o and p, but the values of
-         * the properties in p are discarded
-         */
-//        sjl.intersection = function (o, p) {
-//            return sjl.restrict(sjl.clone(o), p);
-//        };
-//    }
-
-})(typeof window === 'undefined' ? global : window);
