+++ src/sjl/stdlib/Attributable.js
