+++ src/sjl/stdlib/Extendable.js
