+++ src/sjl/stdlib/Iterator.js
