+++ src/sjl/stdlib/SjlSearchableMap.js
